# Speed up LiverpoolToVK::SurfaceFormat