In Bloodborne this shows up as the function with the very highest cumulative "exclusive time". This is true both in scenes that perform poorly, and scenes that perform well.

I took (approximately) 10s samples using an 8khz sampling profiler.

In the Nightmare Grand Cathedral (looking towards the stairs, at the rest of the level):
- Reduced total time from 757.34ms to 82.61ms (out of ~10000ms).
- Reduced average frame times by 2ms (though according to the graph, the gap may be as big as 9ms every N frames).

[00-savedata.zip](https://github.com/user-attachments/files/18276958/00-savedata.zip)
Before:
![00-before](https://github.com/user-attachments/assets/d0a6547d-ae9a-41d5-b31e-07ef82f838fe)
After:
![00-after](https://github.com/user-attachments/assets/ea18dbd6-3f22-4cf7-a7e2-55056ea21d13)
Before:
![00-before-profile](https://github.com/user-attachments/assets/b0f605e2-b52c-49ae-9ccd-0e9a94847191)
After:
![00-after-profile](https://github.com/user-attachments/assets/5807dc4c-9b4a-49f1-bf07-5d75d5a40824)

In the Hunter's Dream (in the spawn position):
- Reduced the total time from 486.50ms to 53.83ms (out of ~10000ms).
- Average frame times appear to be roughly the same.

[01-savedata.zip](https://github.com/user-attachments/files/18276964/01-savedata.zip)
Before:
![01-before](https://github.com/user-attachments/assets/feef3cbf-92c3-46ee-9acb-28a1c074300f)
After:
![01-after](https://github.com/user-attachments/assets/6a84dc24-8b1c-4df2-9ac9-a60fef599e6c)
Before:
![01-before-profile](https://github.com/user-attachments/assets/e00b13b7-c3ce-48dc-8606-91bc81925d70)
After:
![01-after-profile](https://github.com/user-attachments/assets/dcfe2472-da6e-4b42-a84c-5eb4eb9ca4c9)

These are profiles of the change vs the version currently in the main branch. These improvements also improve things in the `locking` branch. They might improve them even more in that branch, but I didn't bother keeping track of my measurements as well in that branch. I believe this change will still be useful even when that branch is stabilized and merged.

It could be there are other bottlenecks in rendering on the main branch that are preventing this code from being the critical path in places like the Hunter's Dream, where the frame times aren't as high. That might explain why the reduction in call times isn't resulting in a higher frame rate in every scene.