@bigol83 2024-12-30
This improves Resident Evil 4 and Yakuza 0 too by 2 fps.