@squidbus 2024-12-30
This is a nice perf improvement but I would prefer a solution that does not now involve maintaining two different copies of format mappings.