@squidbus 2024-12-31
One suggestion would be if we could bitwise combine data and number formats into an indexed lookup to replace the `find_if`.