@hspir404 2024-12-31
Changed to a lookup table.

It adds a 4kb table (6 bits * 4 bits * 4 bytes per vulkan format), but seems to result in a smaller executable than when the switch statement was there (even though I verified it should only be 975 bytes of code). Confused how that works, but apparently this is a 35MB executable anyway so I guess we're not sweating such things.

It also runs a gnat's hair faster than the switch statement did on my computer.

![after](https://github.com/user-attachments/assets/9a9c192a-2ef5-4bcf-b308-8769d56144d1)

Guess I should do a clang format to tidy this now.